Rename YamlNodeRegistry -> LocalYamlNodeRegistry
Renames the `YamlNodeRegistry` to `LocalYamlNodeRegistry` to prevent
confusion with the incoming `RemoteYamlNodeRegistry` and moves the local
YAML registry into the `splinter::node_registry::yaml::local` module.

Also updates the node registry module to keep its sub-modules private
(except for rest_api) in favor of re-exporting the structs in the
sub-modules.
